+++ packages/atproto_core/lib/src/sessions/auth_token.g.dart
@@ -1,17 +1,33 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
+// ignore_for_file: non_constant_identifier_names
+
 part of 'auth_token.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$AuthTokenImpl _$$AuthTokenImplFromJson(Map<String, dynamic> json) =>
-    _$AuthTokenImpl(
-      scope: $enumDecode(_$AuthScopeEnumMap, json['scope']),
-      subject: json['sub'] as String,
-      expiresAt: dateTimeConverter.fromJson(json['exp'] as int),
-      issuedAt: dateTimeConverter.fromJson(json['iat'] as int),
+_$AuthTokenImpl _$$AuthTokenImplFromJson(Map json) => $checkedCreate(
+      r'_$AuthTokenImpl',
+      json,
+      ($checkedConvert) {
+        final val = _$AuthTokenImpl(
+          scope: $checkedConvert(
+              'scope', (v) => $enumDecode(_$AuthScopeEnumMap, v)),
+          subject: $checkedConvert('sub', (v) => v as String),
+          expiresAt: $checkedConvert(
+              'exp', (v) => dateTimeConverter.fromJson(v as int)),
+          issuedAt: $checkedConvert(
+              'iat', (v) => dateTimeConverter.fromJson(v as int)),
+        );
+        return val;
+      },
+      fieldKeyMap: const {
+        'subject': 'sub',
+        'expiresAt': 'exp',
+        'issuedAt': 'iat'
+      },
     );
 
 Map<String, dynamic> _$$AuthTokenImplToJson(_$AuthTokenImpl instance) =>