+++ packages/atproto_core/lib/src/types/blob.g.dart
@@ -1,16 +1,27 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
+// ignore_for_file: non_constant_identifier_names
+
 part of 'blob.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$BlobImpl _$$BlobImplFromJson(Map<String, dynamic> json) => _$BlobImpl(
-      type: json[r'$type'] as String? ?? 'blob',
-      mimeType: json['mimeType'] as String,
-      size: json['size'] as int,
-      ref: BlobRef.fromJson(json['ref'] as Map<String, dynamic>),
+_$BlobImpl _$$BlobImplFromJson(Map json) => $checkedCreate(
+      r'_$BlobImpl',
+      json,
+      ($checkedConvert) {
+        final val = _$BlobImpl(
+          type: $checkedConvert(r'$type', (v) => v as String? ?? 'blob'),
+          mimeType: $checkedConvert('mimeType', (v) => v as String),
+          size: $checkedConvert('size', (v) => v as int),
+          ref: $checkedConvert('ref',
+              (v) => BlobRef.fromJson(Map<String, Object?>.from(v as Map))),
+        );
+        return val;
+      },
+      fieldKeyMap: const {'type': r'$type'},
     );
 
 Map<String, dynamic> _$$BlobImplToJson(_$BlobImpl instance) =>
@@ -18,5 +29,5 @@ Map<String, dynamic> _$$BlobImplToJson(_$BlobImpl instance) =>
       r'$type': instance.type,
       'mimeType': instance.mimeType,
       'size': instance.size,
-      'ref': instance.ref,
+      'ref': instance.ref.toJson(),
     };