+++ packages/atproto_core/lib/src/types/blob_ref.g.dart
@@ -1,14 +1,23 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
+// ignore_for_file: non_constant_identifier_names
+
 part of 'blob_ref.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$BlobRefImpl _$$BlobRefImplFromJson(Map<String, dynamic> json) =>
-    _$BlobRefImpl(
-      link: json[r'$link'] as String,
+_$BlobRefImpl _$$BlobRefImplFromJson(Map json) => $checkedCreate(
+      r'_$BlobRefImpl',
+      json,
+      ($checkedConvert) {
+        final val = _$BlobRefImpl(
+          link: $checkedConvert(r'$link', (v) => v as String),
+        );
+        return val;
+      },
+      fieldKeyMap: const {'link': r'$link'},
     );
 
 Map<String, dynamic> _$$BlobRefImplToJson(_$BlobRefImpl instance) =>